move chat view component to new file
